uintptr_t Answers

+++ src/core/FtpFileSystem.cpp
@@ -834,7 +834,7 @@ bool TFTPFileSystem::ConfirmOverwrite(UnicodeString & FileName,
     // retry = "resume"
     // all = "yes to newer"
     // ignore = "rename"
-    int Answers = qaYes | qaNo | qaCancel | qaYesToAll | qaNoToAll | qaAll | qaIgnore;
+    uintptr_t Answers = qaYes | qaNo | qaCancel | qaYesToAll | qaNoToAll | qaAll | qaIgnore;
     if (CanResume)
     {
       Answers |= qaRetry;
